Renaming lib alba, centralised -> centralized
