+++ benches/centralized_repetition.rs
@@ -3,7 +3,7 @@ use rand_chacha::ChaCha20Rng;
 use rand_core::SeedableRng;
 use std::time::Duration;
 
-use caledonia::centralised::Proof;
+use alba::centralized::Proof;
 
 pub mod criterion_helpers;
 pub mod utils;
@@ -15,7 +15,7 @@ fn prepetitions(
     n_p: &[usize],
     n_f: &[usize],
 ) {
-    let mut group = c.benchmark_group("Centralised".to_string());
+    let mut group = c.benchmark_group("centralized".to_string());
 
     fn prove_repetitions(
         l: usize,
@@ -30,7 +30,7 @@ fn prepetitions(
         for _ in 0..n {
             // Setup
             let (mut dataset, bench_setup) =
-                utils::setup_centralised_wrapper(&mut rng, l, sp, np, nf);
+                utils::setup_centralized_wrapper(&mut rng, l, sp, np, nf);
             dataset.truncate(truncate_size);
             // Bench
             black_box({
