+++ benches/centralized_step.rs
@@ -3,7 +3,7 @@ use rand_chacha::ChaCha20Rng;
 use rand_core::SeedableRng;
 use std::time::Duration;
 
-use caledonia::centralised::Proof;
+use alba::centralized::Proof;
 
 pub mod criterion_helpers;
 pub mod utils;
@@ -15,7 +15,7 @@ fn psteps(
     n_p: &[usize],
     n_f: &[usize],
 ) {
-    let mut group = c.benchmark_group("Centralised".to_string());
+    let mut group = c.benchmark_group("centralized".to_string());
 
     fn prove_steps(l: usize, sp: usize, np: usize, nf: usize, truncate_size: usize, n: u64) -> u64 {
         let mut rng = ChaCha20Rng::from_entropy();
@@ -23,7 +23,7 @@ fn psteps(
         for _ in 0..n {
             // Setup
             let (mut dataset, bench_setup) =
-                utils::setup_centralised_wrapper(&mut rng, l, sp, np, nf);
+                utils::setup_centralized_wrapper(&mut rng, l, sp, np, nf);
             dataset.truncate(truncate_size);
             // Bench
             black_box({
